Created a build script for windows
In order to enable building and compiling for windows I have constructed a bash script(will require cygwin or similar to run on Windows). This script downloads and compiles all of the required libraries in the appropriate way and enable the next step which is pip installing the package on Windows.

ISSUE-156